Use the Gradle Wrapper for build
Based on https://github.com/qupath/qupath-extension-template

While there may well be arguments for sticking with Maven, switching to Gradle makes it easier to develop and debug within an IDE (at least IntelliJ) since QuPath itself is a Gradle project. It also means the instructions from the template repo can be used.

(This might not be necessary, but I couldn't figure out a setup that allowed me to import the Maven-based extension as a Gradle subproject)